+++ CHANGELOG.md
@@ -7,9 +7,11 @@
 ### Added
 
 - [x] Migrate storage
-- [ ] Persist muted state
+- [x] Persist muted state
 - [ ] Show tab count in folded state
 
 ### Fixed
 
 - [x] Listen to tabs changes for the app header tabs preview
+- [x] Reset `autoloadId` when the corresponding tab is removed from the collection
+- [x] Reset pinner data when migration fails & save a backup of invalid data